refactor: replace strum with serde_plain
The `strum` crate was used only for its enum-from-string capabilities.
To be precise, we needed only FromStr and Display implementations for a
few enums. Those can easily be done manually, but due to poor choises we
ended up with `strum`.

`serde_plain` is very lightweight and provides the needed bare minimum,
without addind any extra dependencies.