chore: add examples for signin & signout
The sign in example can be used to easily get an auth token for the
default user on the account:

$ cargo run -p plex-api --example get-token

It will ask for the OTP if your account has 2FA enabled.

The second example can be used to remove an existing device (e.g. a
client you previously authenticated with the `get-token` example), using
an auth token:

$ cargo run -p plex-api --example signout