refactor(features): generate Feature enum from API
Plex semi-regularly introduces new features, and it's hard to keep up
with them manually. Luckily for me, they provide an HTTP endpoint that
responds with all the available features. This commits adds the means for
the automatic code-generation of the Feature enum based on that API.

The next step is to run this generation on the regular basis and
automatically create a PR with the changes.