test: Add an xtask command to start up a plex server to make changes to the test data.
The task simplifies making changes to the test data. It copies the
existing stub data to the data directory and runs the oldest supported
plex server against it, the assumption is that newer versions may make
backwards incompatible changes to the database. The user can then
connect to the web interface to make modifications as they desire. When
done the stub data is overwritten with the new data.

Adds to the update_pms workflow to automatically set the oldest tag
for the task.

One additional change here is maintaining symlinks when copying plex
data around.

The database changes in this are just those from running the newer plex
server, no actual changes to the library have been made.