test: try catching Plex crashed in transcoding tests
Plex running in a Docker Conatiner (and maybe in other ways too) can
sometimes crash when you start and stop a transcoding session too often.

There's no clear limits or indications of the issue, so we'll just try
ensuring that the server is alive before a tests starts and after it
ends.

At least Plex automatically restarts inside a container in case of a
crash...

Closes: #541